ctray:  Updated build.bat so that it is buildable within this directory structure.
I also addressed the issue that you can build it from whichever directory you wish, instead of needing to be within the same directory as build.bat.

settings.ctray will be copied if the source file is newer to the build directory.

Finally, added the /nologo option for both calls to surpress Microsoft's logo when building.

The file still assumes that vcvarsall.bat has been run before.  At the moment I do not want to add a call for there are many different versions of Visual Studio, and I know I probably am not using one that others use.