ctray:  Added new feature:  Always on top.
Every 30 seconds we will check to see if both of our overlays are the topmost visible window.  If not, we will bump it to the top.  This is to the address what Casey runs into on Handmade Hero where the overlay sometimes disappears.

A hotkey was also added so you can initiate the bump yourself.

It should be noted that the visibility status of our overlays are not taken into account, and even if they are hidden, we will still bump.  Potentially a fix to address later?

The interval time (30 seconds) is hardcoded.  A todo for later is bring this to our settings file.