ctray:  Added new feature:  Save the day!
I changed settings.ctray file to an ini type, and added two additional settings:  Day and TopmostCheckInSeconds.

TopmostCheckInSeconds is the variable that sets the interval timer for how frequent we check if our overlays are on top.  Set to 0 if you want to disable this check.  Default is 30 seconds.

Day starts at 400 for that is the current episode of Handmade Hero as of 2017-07-19.  Note this field will be updated by the program when you begin a session.  This is to remember the day.

A temporary file will be created to save the settings information first.  So the user needs to have write permissions to the directory.

Added some error messaging if there are problems trying to load/save the settings file.  A system tray notification bubble will appear with the Windows error message.

Known issues:  If you run ctray.exe not from the build directory, it won't be able to find settings.ctray.  In other words, if the working directory is not the build directory, the defaults for the display will be used.  If you begin a session, a new settings.ctray file will be created in the current working directory.