ctray:  Some minor corrections after I re-read my code.
The wording in a comment regarding referring to Raymond Chen's blog was incorrect.  It's not that we cannot use FORMAT_MESSAGE_IGNORE_INSERTS, but rather we cannot use inserts.

Since I know the sizes of the buffers, changed some strcpy calls to strncpy.

Added some missing comments of code that I felt needed to put the "why", like why the file name has ./ in front (or else the WinAPI calls I make will search the Windows directory, and not the current).