ctray:  submitted Handmade hero day 401
Pulled changes from https://github.com/brian-hmn/ctray/tree/master/code @ 4918c01

Every 15 seconds we will check to see if both of our overlays are the topmost visible window.  If not, we will bump it to the top.  This is to the address what Casey runs into on Handmade Hero where the overlay sometimes disappears.

It should be noted that the visibility status of our overlays are not taken into account, and even if they are hidden, we will still bump.  Potentially a fix to address later?

The interval time (15 seconds) is hardcoded.  A todo for later is bring this to our settings file.